requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
sqlalchemy==2.0.23
pydantic==2.10.3
pydantic-settings==2.6.1
openai==1.3.7
pinecone-client==5.0.1
python-dotenv==1.0.0
streamlit==1.28.2
